Csl::AuthorName - move AuthorName into a namespace to avoid conflicts with ScienceWire::AuthorName
